Fix failing tests and react to code review
Fix some failing tests, the mapping of hosts/ports was broken by the
earlier changes.

Move some core around, as suggested during code review.
